pull some LSP channel code out of proc macro for better debugging
